% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DT_country_data}
\alias{DT_country_data}
\title{World Bank country level data}
\format{
\subsection{\code{who}}{

A data.table
\describe{
\item{iso3c}{ISO country codes}
\item{country }{Country names}
\item{population}{population}
\item{population_growth_percent}{population growth percent}
\item{labor_force_participation_percent}{Percentage of working age people (15-64) who work}
\item{working_age_percent}{percetange of the population which is in working age (15-64)}
...
}
}
}
\source{
World bank indicator API
}
\usage{
DT_country_data
}
\description{
World Bank country level data
}
\keyword{datasets}
