% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DT_population}
\alias{DT_population}
\title{ILO population data}
\format{
\subsection{\code{who}}{

A data.table
\describe{
\item{ref_area}{ISO country codes}
\item{reference_area}{Country names}
\item{population}{Population}
\item{year}{Year}
...
}
}
}
\source{
https://db.nomics.world/ ILO
}
\usage{
DT_population
}
\description{
Population data by country
}
\keyword{datasets}
