% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DT_population_growth}
\alias{DT_population_growth}
\title{ILO population growth data}
\format{
\subsection{\code{who}}{

A data.table
\describe{
\item{ref_area}{ISO country codes}
\item{reference_area}{Country names}
\item{population_growth}{Population growth for the given year}
\item{year}{Year}
...
}
}
}
\source{
https://db.nomics.world/ ILO
}
\usage{
DT_population_growth
}
\description{
Population growth data by country
}
\keyword{datasets}
