% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_government_expenditure.R
\name{model_government_expenditure}
\alias{model_government_expenditure}
\title{model_government_expenditure}
\usage{
model_government_expenditure(num_years, initial_expenditure, inflation)
}
\arguments{
\item{num_years}{integer number of years to simulate}

\item{initial_expenditure}{expenditure in year 1}

\item{inflation}{inflation rate}
}
\value{
vector of length num_years
}
\description{
model_government_expenditure
}
\examples{
model_government_expenditure(5, 1000, 0.01)
}
