% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_interest.R
\name{model_interest}
\alias{model_interest}
\title{model_interest}
\usage{
model_interest(yearly_contributions, interest_rate)
}
\arguments{
\item{yearly_contributions}{numeric vector of yearly contributions going into the system. from year 1 to year n}

\item{interest_rate}{numeric scalar interest_rate}
}
\value{
numeric vector of interest generated in each year
}
\description{
model_interest
}
\examples{
model_interest(c(1000, 1250, 1600), 0.06)
}
