% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_population_workforce.R
\name{model_population_and_workforce}
\alias{model_population_and_workforce}
\title{model_population_and_workforce}
\usage{
model_population_and_workforce(
  current_population = 5.7e+07,
  pop_growth_current = 0.02,
  percent_working_age = 0.6,
  workforce_participation_rate = 0.42,
  informal_percent_current = 0.8,
  period_in_years = 10
)
}
\arguments{
\item{current_population}{current_population}

\item{pop_growth_current}{population growth rate}

\item{percent_working_age}{percentage of people who are of working age}

\item{workforce_participation_rate}{percentage of working age people who actually work}

\item{informal_percent_current}{percentage of informal workers (out of all workers)}

\item{period_in_years}{number of years to simulate}
}
\value{
a data.table containing year number, population, informal workers etc.
}
\description{
model_population_and_workforce
}
\examples{
model_population_and_workforce()
}
