% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiego.social.security-package.R
\docType{package}
\name{wiego.social.security-package}
\alias{wiego.social.security}
\alias{wiego.social.security-package}
\title{wiego.social.security: Social security cost calculator}
\description{
Shiny application housing a cost calculator for social security subsidies.
}
\author{
\strong{Maintainer}: Jonathan Belke \email{jonathan.belke314@gmail.com}

}
\keyword{internal}
